riwayat yang muncul sudah sesuai idPengguna yang Login

+++ app/src/main/java/com/example/bismillahsipfo/data/model/Gamifikasi.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class Gamifikasi (
     @SerialName("id_gamifikasi") val idGamifikasi: Int,
     @SerialName("level") val level: Int,
-    @SerialName("jumlah_peminjaman_minimal") val jumlahPeminjamanMinimal: Int,
+    @SerialName("jumlah_peminjaman_minimal") val jumlahPeminjamanMinimal: Double,
     @SerialName("tropi") val tropi: String,
     @SerialName("id_voucher") val idVoucher: Int
 )
